+++ paq8pxd.cpp
@@ -125,10 +125,12 @@ inline int max(int a, int b) {return a<b?b:a;}
 #include "prt/sscm.hpp"
 #include "prt/stationarymap.hpp"
 #include "prt/mft.hpp"
+#include "prt/contextmap2.hpp"
+
 #include "models/model.hpp"
 #include "models/sparsematch.hpp"
 #include "models/exe.hpp"
-#include "prt/contextmap2.hpp"
+#include "models/dec.hpp"
 
 /////////////////////// Global context /////////////////////////
 U8 level=DEFAULT_OPTION;  // Compression level 0 to 15
@@ -2117,6 +2119,7 @@ public:
     return res;
   }
 };
+/*
 bool CAlpha2(U32 ins) {
    static const int op10[22]= {0x40,0x00,0x02,0x49,0x09,0x0b,0x0f,0x12,0x1b,0x1d,0x60,0x20,0x22,0x69,0x29,0x2b,0x2d,0x32,0x3b,0x3d,0x4d,0x6d};
    static const int op11[16]= {0x00,0x08,0x14,0x16,0x20,0x24,0x26,0x28,0x40,0x44,0x46,0x48,0x61,0x64,0x66,0x6c};
@@ -2238,7 +2241,7 @@ bool CAlpha2(U32 ins) {
     }
     if (opc>=0x1d && opc<=0x3f) return true;
     return false;
-}
+}*/
 //////////////////////////// Models //////////////////////////////
 
 // All of the models below take a Mixer as a parameter and write
@@ -3423,11 +3426,7 @@ int matchModel2::p(Mixer &m,int val1=0,int val2=0) {
 //Based on XWRT 3.2 (29.10.2007) - XML compressor by P.Skibinski, inikep@gmail.com
 #include "wrton.cpp"
 //////////////////////////// wordModel /////////////////////////
-#define bswap(x) \
-+   ((((x) & 0xff000000) >> 24) | \
-+    (((x) & 0x00ff0000) >>  8) | \
-+    (((x) & 0x0000ff00) <<  8) | \
-+    (((x) & 0x000000ff) << 24))
+
 #define TAB 0x09
 #define NEW_LINE 0x0A
 #define CARRIAGE_RETURN 0x0D
@@ -8192,828 +8191,7 @@ int p(Mixer& m,int val1=0,int val2=0){
  }
 };
 
-template<class T>
-class RingBuffer1 {
-private:
-    Array<T> b;
-    uint32_t offset {0}; /**< Number of input bytes in buffer (not wrapped), will be masked when used for indexing */
-    uint32_t mask;
 
-public:
-    /**
-     * Creates an array of @ref size bytes (must be a power of 2).
-     * @param size number of bytes in array
-     */
-    explicit RingBuffer1(const uint32_t size = 0) : b(size), mask(size - 1) {
-#ifdef VERBOSE
-      printf("Created RingBuffer with size = %d\n", size);
-#endif
-      assert(isPowerOf2(size));
-    }
-
-    void setSize(uint32_t newSize) {
-      assert(newSize > 0 && isPowerOf2(newSize));
-      b.resize(newSize);
-      offset = 0;
-      mask = newSize - 1;
-    }
-
-     auto getpos() const -> uint32_t {
-      return offset;
-    }
-
-    void fill(const T B) {
-      const auto n = (uint32_t) b.size();
-      for( uint32_t i = 0; i < n; i++ ) {
-        b[i] = B;
-      }
-    }
-
-    void add(const T B) {
-      b[offset & mask] = B;
-      offset++;
-    }
-
-    /**
-     * Returns a reference to the i'th byte with wrap (no out of bounds).
-     * @param i
-     * @return
-     */
-    auto operator[](const uint32_t i) const -> T {
-      return b[i & mask];
-    }
-
-    void set(const uint32_t i, const T B) {
-      b[i & mask] = B;
-    }
-
-    /**
-     * Returns i'th byte back from pos (i>0) with wrap (no out of bounds)
-     * @param i
-     * @return
-     */
-    auto operator()(const uint32_t i) const -> T {
-      //assert(i!=0);
-      return b[(offset - i) & mask];
-    }
-
-    void reset() {
-      fill(0);
-      offset = 0;
-    }
-
-    /**
-     * @return the size of the RingBuffer
-     */
-    auto size() -> uint32_t {
-      return (uint32_t) b.size();
-    }
-
-    void copyTo(RingBuffer1 &dst) {
-      dst.setSize(size());
-      dst.offset = offset;
-      auto n = (uint32_t) b.size();
-      for( uint32_t i = 0; i < n; i++ ) {
-        dst.b[i] = b[i];
-      }
-    }
-};
-
-class DECAlpha {
-public:
-  enum InstructionFormat { Bra, F_P, Mem, Mfc, Mbr, Opr, Pcd, Nop };
-  static std::uint32_t const op10[];
-  static std::uint32_t const op11[];
-  static std::uint32_t const op12[];
-  static std::uint32_t const op13[];
-  static std::uint32_t const op14[];
-  static std::uint32_t const op15a[];
-  static std::uint32_t const op15b[];
-  static std::uint32_t const op16a[];
-  static std::uint32_t const op16b[];
-  static std::uint32_t const op17[];
-  static std::uint32_t const op18[];
-  static std::uint32_t const op1C[];
-  static InstructionFormat const formats[];
-  static std::uint8_t const rel_op[];
-  static bool IsValidInstruction(std::uint32_t const instruction);
-  static void Shuffle(std::uint32_t& instruction);
-  static void Unshuffle(std::uint32_t& instruction);
-};
-
-std::uint32_t const DECAlpha::op10[] = { 0x40u,0x00u,0x02u,0x49u,0x09u,0x0Bu,0x0Fu,0x12u,0x1Bu,0x1Du,0x60u,0x20u,0x22u,0x69u,0x29u,0x2Bu,0x2Du,0x32u,0x3Bu,0x3Du,0x4Du,0x6Du };
-std::uint32_t const DECAlpha::op11[] = { 0x00u,0x08u,0x14u,0x16u,0x20u,0x24u,0x26u,0x28u,0x40u,0x44u,0x46u,0x48u,0x61u,0x64u,0x66u,0x6Cu };
-std::uint32_t const DECAlpha::op12[] = { 0x02u,0x06u,0x0Bu,0x12u,0x16u,0x1Bu,0x22u,0x26u,0x2Bu,0x30u,0x31u,0x32u,0x34u,0x36u,0x39u,0x3Bu,0x3Cu,0x52u,0x57u,0x5Au,0x62u,0x67u,0x6Au,0x72u,0x77u,0x7Au };
-std::uint32_t const DECAlpha::op13[] = { 0x40u,0x00u,0x60u,0x20u,0x30u };
-std::uint32_t const DECAlpha::op14[] = { 0x004u,0x00Au,0x08Au,0x10Au,0x18Au,0x40Au,0x48Au,0x50Au,0x58Au,0x00Bu,0x04Bu,0x08Bu,0x0CBu,0x10Bu,
-                                         0x14Bu,0x18BU,0x1CBu,0x50Bu,0x54Bu,0x58Bu,0x5CBu,0x70Bu,0x74Bu,0x78Bu,0x7CBu,0x014u,0x024u,0x02Au,
-                                         0x0AAu,0x12Au,0x1AAu,0x42Au,0x4AAu,0x52Au,0x5AAu,0x02Bu,0x06Bu,0x0ABu,0x0EBu,0x12Bu,0x16Bu,0x1ABu,
-                                         0x1EBu,0x52Bu,0x56Bu,0x5ABu,0x5EBu,0x72Bu,0x76Bu,0x7ABu,0x7EBu };
-std::uint32_t const DECAlpha::op15a[] = { 0x0A5u,0x4A5u,0x0A6u,0x4A6u,0x0A7u,0x4A7u,0x03Cu,0x0BCu,0x03Eu,0x0BEu };
-std::uint32_t const DECAlpha::op15b[] = { 0x000u,0x001u,0x002u,0x003u,0x01Eu,0x020u,0x021u,0x022u,0x023u,0x02Cu,0x02Du,0x02Fu };
-std::uint32_t const DECAlpha::op16a[] = { 0x0A4u,0x5A4u,0x0A5u,0x5A5u,0x0A6u,0x5A6u,0x0A7u,0x5A7u,0x2ACu,0x6ACu };
-std::uint32_t const DECAlpha::op16b[] = { 0x00u,0x01u,0x02u,0x03u,0x20u,0x21u,0x22u,0x23u,0x2Cu,0x2Fu };
-std::uint32_t const DECAlpha::op17[] = { 0x010u,0x020u,0x021u,0x022u,0x024u,0x025u,0x02Au,0x02Bu,0x02Cu,0x02Du,0x02Eu,0x02Fu,0x030u,0x130u,0x530u };
-std::uint32_t const DECAlpha::op18[] = { 0x0000u,0x0400u,0x4000u,0x4400u,0x8000u,0xA000u,0xC000u,0xE000u,0xE800u,0xF000u,0xF800u,0xFC00u };
-std::uint32_t const DECAlpha::op1C[] = { 0x00u,0x01u,0x30u,0x31u,0x32u,0x33u,0x34u,0x35u,0x36u,0x37u,0x38u,0x39u,0x3Au,0x3Bu,0x3Cu,0x3Du,0x3Eu,0x3Fu,0x70u,0x78u };
-
-DECAlpha::InstructionFormat const DECAlpha::formats[] = {
-  DECAlpha::Pcd, DECAlpha::Nop, DECAlpha::Nop, DECAlpha::Nop, DECAlpha::Nop, DECAlpha::Nop, DECAlpha::Nop, DECAlpha::Nop, // 00..07
-  DECAlpha::Mem, DECAlpha::Mem, DECAlpha::Mem, DECAlpha::Mem, DECAlpha::Mem, DECAlpha::Mem, DECAlpha::Mem, DECAlpha::Mem, // 08..0F
-  DECAlpha::Opr, DECAlpha::Opr, DECAlpha::Opr, DECAlpha::Opr, DECAlpha::F_P, DECAlpha::F_P, DECAlpha::F_P, DECAlpha::F_P, // 10..17
-  DECAlpha::Mfc, DECAlpha::Pcd, DECAlpha::Mbr, DECAlpha::Pcd, DECAlpha::Opr, DECAlpha::Pcd, DECAlpha::Pcd, DECAlpha::Pcd, // 18..1F
-  DECAlpha::Mem, DECAlpha::Mem, DECAlpha::Mem, DECAlpha::Mem, DECAlpha::Mem, DECAlpha::Mem, DECAlpha::Mem, DECAlpha::Mem, // 20..27
-  DECAlpha::Mem, DECAlpha::Mem, DECAlpha::Mem, DECAlpha::Mem, DECAlpha::Mem, DECAlpha::Mem, DECAlpha::Mem, DECAlpha::Mem, // 28..2F
-  DECAlpha::Bra, DECAlpha::Bra, DECAlpha::Bra, DECAlpha::Bra, DECAlpha::Mbr, DECAlpha::Bra, DECAlpha::Bra, DECAlpha::Bra, // 30..37
-  DECAlpha::Bra, DECAlpha::Bra, DECAlpha::Bra, DECAlpha::Bra, DECAlpha::Bra, DECAlpha::Bra, DECAlpha::Bra, DECAlpha::Bra  // 38..3F
-};
-
-/*
-Bra: 15 opcodes
-F_P:  4 opcodes, 108 functions
-Mem: 24 opcodes
-Mfc:  1 opcode
-Mbr:  2 opcodes
-Opr:  5 opcodes,  89 functions
-Pcd:  6 opcodes
-Nop:  7 opcodes
-*/
-std::uint8_t const DECAlpha::rel_op[] = {
-  0x00u, 0x00u, 0x01u, 0x02u, 0x03u, 0x04u, 0x05u, 0x06u, // 00..07
-  0x00u, 0x01u, 0x02u, 0x03u, 0x04u, 0x05u, 0x06u, 0x07u, // 08..0F
-  0x00u, 0x01u, 0x02u, 0x03u, 0x00u, 0x01u, 0x02u, 0x03u, // 10..17
-  0x00u, 0x01u, 0x00u, 0x02u, 0x04u, 0x03u, 0x04u, 0x05u, // 18..1F
-  0x08u, 0x09u, 0x0Au, 0x0Bu, 0x0Cu, 0x0Du, 0x0Eu, 0x0Fu, // 20..27
-  0x10u, 0x11u, 0x12u, 0x13u, 0x14u, 0x15u, 0x16u, 0x17u, // 28..2F
-  0x00u, 0x01u, 0x02u, 0x03u, 0x01u, 0x04u, 0x05u, 0x06u, // 30..37
-  0x07u, 0x08u, 0x09u, 0x0Au, 0x0Bu, 0x0Cu, 0x0Du, 0x0Eu  // 38..3F
-};
-#include <algorithm>
-bool DECAlpha::IsValidInstruction(std::uint32_t const instruction) {
-  std::uint32_t opcode = instruction >> 26u, function = 0u;
-  switch (opcode) {
-    case 0: { // PAL
-      function = instruction & 0x1FFFFFFFu;
-      return !(((function > 0x3Fu) && (function < 0x80u)) || (function > 0xBFu));
-    }
-    case 0x10u: { // INTA
-      function = (instruction >> 5u) & 0x7Fu;
-      return std::find(std::begin(op10), std::end(op10), function) != std::end(op10);
-    }
-    case 0x11u: { // INTL
-      function = (instruction >> 5u) & 0x7Fu;
-      return std::find(std::begin(op11), std::end(op11), function) != std::end(op11);
-    }
-    case 0x12u: { // INTS
-      function = (instruction >> 5u) & 0x7Fu;
-      return std::find(std::begin(op12), std::end(op12), function) != std::end(op12);
-    }
-    case 0x13u: { // INTM
-      function = (instruction >> 5u) & 0x7Fu;
-      return std::find(std::begin(op13), std::end(op13), function) != std::end(op13);
-    }
-    case 0x14u: { // ITFP
-      function = (instruction >> 5u) & 0x7FFu;
-      return std::find(std::begin(op14), std::end(op14), function) != std::end(op14);
-    }
-    case 0x15u: { // FLTV
-      function = (instruction >> 5u) & 0x7FFu;
-      if (std::find(std::begin(op15a), std::end(op15a), function) != std::end(op15a))
-        return true;
-      if ((function & 0x200u) != 0u)
-        return false;
-      function &= 0x7Fu;
-      return std::find(std::begin(op15b), std::end(op15b), function) != std::end(op15b);
-    }
-    case 0x16u: { // FLTI
-      function = (instruction >> 5u) & 0x7FFu;
-      if (std::find(std::begin(op16a), std::end(op16a), function) != std::end(op16a))
-        return true;
-      if (((function & 0x600u) == 0x200u) || ((function & 0x500u) == 0x400u))
-        return false;
-      opcode = function & 0x3Fu;
-      if (std::find(std::begin(op16b), std::end(op16b), opcode) != std::end(op16b))
-        return true;
-      if ((opcode == 0x3Cu) || (opcode == 0x3Eu))
-        return (function & 0x300u) != 0x100u;
-      return false;
-    }
-    case 0x17u: { // FLTL
-      function = (instruction >> 5u) & 0x7FFu;
-      return std::find(std::begin(op17), std::end(op17), function) != std::end(op17);
-    }
-    case 0x18u: {
-      function = instruction & 0xFFFFu;
-      return std::find(std::begin(op18), std::end(op18), function) != std::end(op18);
-    }
-    case 0x1Cu: { // FPTI
-      function = (instruction >> 5u) & 0x7Fu;
-      return std::find(std::begin(op1C), std::end(op1C), function) != std::end(op1C);
-    }
-    default: {
-      if ((opcode >= 1u) && (opcode <= 7u))
-        return false;
-      if (((opcode >= 0x08u) && (opcode <= 0x0Fu)) || ((opcode >= 0x19u) && (opcode <= 0x3Fu)))
-        return true;
-    }
-  }
-  return false;
-}
-
-void DECAlpha::Shuffle(std::uint32_t& instruction) {
-  std::uint32_t const opcode = instruction >> 26u;
-  switch (DECAlpha::formats[opcode]) {
-    case DECAlpha::F_P: { // op, ra, rb, func, rc => op, func, ra, rb, rc
-      instruction = (instruction & 0xFC000000u) | 
-                    ((instruction & 0xFFE0u) << 10u) |
-                    ((instruction >> 11u) & 0x7FE0u) |
-                    (instruction & 0x1Fu);
-      break;
-    }
-    case DECAlpha::Mfc: { // op, ra, rb, func => op, func, ra, rb
-      instruction = (instruction & 0xFC000000u) |
-                    ((instruction & 0xFFFFu) << 10u) |
-                    ((instruction >> 16u) & 0x3FFu);
-      break;
-    }
-    case DECAlpha::Opr: { // op, ra, {(rb, unused), (literal)}, bit, func, rc => op, bit, func, ra, {(rb, unused), (literal)}, rc
-      instruction = (instruction & 0xFC000000u) |
-                    ((instruction << 13u) & 0x3FC0000u) |
-                    ((instruction >> 8u) & 0x3FFE0u) |
-                    (instruction & 0x1Fu);
-      break;
-    }
-  }
-  instruction = bswap(instruction);
-}
-
-void DECAlpha::Unshuffle(std::uint32_t& instruction) {
-  instruction = bswap(instruction);
-  std::uint32_t const opcode = instruction >> 26u;
-  switch (DECAlpha::formats[opcode]) {
-    case DECAlpha::F_P: { // op, func, ra, rb, rc => op, ra, rb, func, rc
-      instruction = (instruction & 0xFC000000u) |
-                    ((instruction >> 10u) & 0xFFE0u) |
-                    ((instruction & 0x7FE0u) << 11u) |
-                    (instruction & 0x1Fu);
-      break;
-    }
-    case DECAlpha::Mfc: { // op, func, ra, rb => op, ra, rb, func
-      instruction = (instruction & 0xFC000000u) |
-                    ((instruction >> 10u) & 0xFFFFu) |
-                    ((instruction & 0x3FFu) << 16u);
-      break;
-    }
-    case DECAlpha::Opr: { // op, bit, func, ra, {(rb, unused), (literal)}, rc => op, ra, {(rb, unused), (literal)}, bit, func, rc
-      instruction = (instruction & 0xFC000000u) |
-                    ((instruction & 0x3FC0000u) >> 13u) |
-                    ((instruction & 0x3FFE0u) << 8u) |
-                    (instruction & 0x1Fu);
-      break;
-    }
-  }
-}
-class decModel1: public Model {
-public:  
-  BlockData& x;
-  Buf& buf;
-    enum State {
-    OpCode,
-    Bra_Ra,
-    Bra_Displacement,
-    F_P_Function,
-    F_P_Ra,
-    F_P_Rb,
-    F_P_Rc,
-    Mem_Ra,
-    Mem_Rb,
-    Mem_Displacement,
-    Mfc_Function,
-    Mfc_Ra,
-    Mfc_Rb,
-    Mbr_Ra,
-    Mbr_Rb,
-    Mbr_Displacement,
-    Opr_Bit,
-    Opr_Function,
-    Opr_Ra,
-    Opr_Rb,
-    Opr_Unused,
-    Opr_Literal,
-    Opr_Rc,
-    Pcd_Function,
-    Nop_Skip,
-    Count
-  };
-  struct Instruction {
-    std::uint32_t Function, Displacement;
-    std::uint8_t Opcode, Bit, Ra, Rb, Rc, Literal, relOpcode, Format;
-  };
-  
-  static constexpr std::uint32_t nMaps = 11u;
-   std::uint32_t maps_mask[nMaps-1] = { 0x6FC3FFu, 0x6F0387u, 0x4E0383u, 0x440183u, 0x181u, 0x181u, 0x81u, 0x1u, 0x1u, 0x1u };
- private:
-  RingBuffer1<Instruction> cache;
-  State state;
-  Instruction op;
-  std::uint32_t count;
-  std::uint8_t lastRc;
-  Instruction last[8];
-  IndirectMap maps0[State::Count], maps1[18], maps2[12], maps3[9], maps4[6], maps5[3], maps6[3], maps7[2], maps8[1], maps9[1], maps10[1];
-  IndirectMap* const maps[nMaps - 1] = { &maps1[0], &maps2[0], &maps3[0], &maps4[0], &maps5[0], &maps6[0], &maps7[0], &maps8[0], &maps9[0], &maps10[0] };
-
-
-           std::int32_t map_state(std::uint32_t const map, State const state) {
-    std::int32_t r = -1;
-    if ((map < nMaps) && (((maps_mask[map] >> state) & 1u) != 0u))     
-      for (std::int32_t i = state; i >= 0; r += (maps_mask[map] >> i) & 1, i--);
-    return r;
-  }
-  public:
-  decModel1(BlockData& bd,U32 val=0):x(bd),buf(bd.buf),
-   cache(8), op{0}, state(OpCode), count(0u), lastRc(0xFF), last{0},
-    maps0 {
-      { 1, 6, 256, 255}, // OpCode
-      { 1, 5, 256, 255}, // Bra_Ra
-      { 2, 7, 256, 255}, // Bra_Displacement
-      { 2, 3, 256, 255}, // F_P_Function
-      { 1, 5, 256, 255}, // F_P_Ra
-      { 1, 5, 256, 255}, // F_P_Rb
-      { 1, 5, 256, 255}, // F_P_Rc
-      { 1, 5, 256, 255}, // Mem_Ra
-      { 1, 5, 256, 255}, // Mem_Rb
-      { 2, 4, 256, 255}, // Mem_Displacement
-      { 2, 4, 512, 255}, // Mfc_Function
-      { 1, 5, 512, 255}, // Mfc_Ra
-      { 1, 5, 512, 255}, // Mfc_Rb
-      { 1, 5, 512, 255}, // Mbr_Ra
-      { 1, 5, 256, 255}, // Mbr_Rb
-      { 2, 4, 256, 255}, // Mbr_Displacement
-      { 1, 1, 256, 255}, // Opr_Bit
-      { 1, 7, 256, 255}, // Opr_Function
-      { 1, 5, 256, 255}, // Opr_Ra
-      { 1, 5, 256, 255}, // Opr_Rb
-      { 1, 3, 512, 255}, // Opr_Unused
-      { 1, 8, 256, 255}, // Opr_Literal
-      { 1, 5, 256, 255}, // Opr_Rc
-      { 4, 3, 512, 255}, // Pcd_Function
-      { 4, 3, 512, 255}  // Nop_Skip
-    },
-    maps1 {
-      { 11, 6, 256, 255}, // OpCode
-      {  5, 5, 256, 255}, // Bra_Ra
-      { 14, 7, 256, 255}, // Bra_Displacement
-      { 14, 3, 256, 255}, // F_P_Function
-      {  5, 5, 256, 255}, // F_P_Ra
-      { 10, 5, 256, 255}, // F_P_Rb
-      { 10, 5, 256, 255}, // F_P_Rc
-      {  5, 5, 256, 255}, // Mem_Ra
-      {  5, 5, 256, 255}, // Mem_Rb
-      { 16, 4, 256, 255}, // Mem_Displacement
-      {  1, 5, 256, 255}, // Mbr_Rb
-      {  3, 4, 256, 255}, // Mbr_Displacement
-      {  3, 1, 256, 255}, // Opr_Bit
-      {  4, 7, 256, 255}, // Opr_Function
-      {  5, 5, 256, 255}, // Opr_Ra
-      { 10, 5, 256, 255}, // Opr_Rb
-      { 10, 8, 256, 255}, // Opr_Literal
-      {  6, 5, 256, 255}  // Opr_Rc
-    },
-    maps2 {
-      { 16, 6, 256, 255}, // OpCode
-      { 10, 5, 256, 255}, // Bra_Ra
-      { 15, 7, 256, 255}, // Bra_Displacement
-      { 10, 5, 256, 255}, // Mem_Ra
-      { 10, 5, 256, 255}, // Mem_Rb
-      { 18, 4, 256, 255}, // Mem_Displacement
-      { 11, 1, 256, 255}, // Opr_Bit
-      { 16, 7, 256, 255}, // Opr_Function
-      {  5, 5, 256, 255}, // Opr_Ra
-      { 10, 5, 256, 255}, // Opr_Rb
-      { 14, 8, 256, 255}, // Opr_Literal
-      { 16, 5, 256, 255}  // Opr_Rc
-    },
-    maps3{
-      { 16, 6, 256, 255}, // OpCode
-      { 16, 5, 256, 255}, // Bra_Ra
-      { 15, 5, 256, 255}, // Mem_Ra
-      { 15, 5, 256, 255}, // Mem_Rb
-      { 18, 4, 256, 255}, // Mem_Displacement
-      { 16, 7, 256, 255}, // Opr_Function
-      { 10, 5, 256, 255}, // Opr_Ra
-      { 15, 5, 256, 255},  // Opr_Rb
-      { 11, 5, 256, 255}  // Opr_Rc
-    },
-    maps4{
-      { 16, 6, 256, 255}, // OpCode
-      { 16, 5, 256, 255}, // Bra_Ra
-      { 16, 5, 256, 255}, // Mem_Ra
-      { 16, 5, 256, 255}, // Mem_Rb
-      { 16, 5, 256, 255}, // Opr_Ra
-      { 16, 5, 256, 255}  // Opr_Rc
-    },
-    maps5{
-      { 17, 6, 256, 255}, // OpCode
-      { 17, 5, 256, 255}, // Mem_Ra
-      { 16, 5, 256, 255}  // Mem_Rb
-    },
-    maps6{
-      { 17, 6, 256, 255}, // OpCode
-      { 17, 5, 256, 255}, // Mem_Ra
-      { 17, 5, 256, 255}  // Mem_Rb
-    },
-    maps7{
-      { 18, 6, 256, 255}, // OpCode
-      { 18, 5, 256, 255}, // Mem_Ra
-    },
-    maps8{
-      { 18, 6, 256, 255}  // OpCode
-    },
-    maps9{
-      { 17, 6, 256, 255}  // OpCode
-    },
-    maps10{
-      { 17, 6, 256, 255}  // OpCode
-    }
-   
-   {
- }
- int inputs() {return nMaps*2 ;}
- int nets() {return State::Count * 26 + State::Count * 64 + 2048 + 4096 + 4096 + 8192 + 8192 + 8192 + 4096 + 4096 + 4096;;}
-  int netcount() {return 11;}
-int p(Mixer& m,int val1=0,int val2=0){  
-  if ((x.blpos  == 0u) && (x.bpos == 0)) {
-      state = State::OpCode;
-      for (std::uint32_t i = 0u; i < nMaps - 1u; i++) {
-        if (((maps_mask[i] >> State::OpCode) & 1u) != 0u)
-          maps[i][map_state(i, State::OpCode)].set(0u);
-      }
-      count = 0u;
-      cache.fill(op = { 0 });
-    }
-    else {
-      count++;
-      //INJECT_SHARED_y
-      switch (state) {
-        case State::OpCode: {
-          op.Opcode += op.Opcode + x.y;
-          if (count == 6u) {
-            op.relOpcode = DECAlpha::rel_op[op.Opcode];
-            switch (op.Format = DECAlpha::formats[op.Opcode]) {
-              case DECAlpha::Bra: {
-                state = State::Bra_Ra;
-                maps1[map_state(0u, State::Bra_Ra)].set_direct(cache(1).Rc);
-                maps2[map_state(1u, State::Bra_Ra)].set_direct((cache(1).Ra << 5u) | cache(1).Rc);
-                maps3[map_state(2u, State::Bra_Ra)].set_direct((cache(1).Opcode << 10u) | (cache(1).Ra << 5u) | cache(1).Rc);
-                maps4[map_state(3u, State::Bra_Ra)].set(hash(cache(1).Ra, cache(2).Ra, cache(3).Ra, cache(4).Ra));
-                break;
-              }
-              case DECAlpha::F_P: {
-                state = State::F_P_Function;
-                maps1[map_state(0u, State::F_P_Function)].set_direct(op.Opcode);
-                break;
-              }
-              case DECAlpha::Mem: {
-                state = State::Mem_Ra;
-                maps1[map_state(0u, State::Mem_Ra)].set_direct(cache(1).Ra);
-                maps2[map_state(1u, State::Mem_Ra)].set_direct((cache(1).Ra << 5u) | cache(2).Ra);
-                maps3[map_state(2u, State::Mem_Ra)].set_direct((cache(1).Ra << 10u) | (cache(2).Ra << 5u) | cache(3).Ra);
-                maps4[map_state(3u, State::Mem_Ra)].set(hash(cache(1).Ra, cache(2).Ra, cache(3).Ra, cache(4).Ra));
-                maps5[map_state(4u, State::Mem_Ra)].set(hash(op.Opcode, cache(1).Opcode, cache(1).Ra, cache(2).Opcode, cache(2).Ra));
-                maps6[map_state(5u, State::Mem_Ra)].set(hash(op.Opcode, cache(1).Ra, cache(2).Ra, cache(2).Rb,  cache(3).Ra, cache(4).Ra));
-                maps7[map_state(6u, State::Mem_Ra)].set(hash(op.Opcode, cache(1).Ra, cache(1).Rb, cache(2).Ra,  cache(3).Ra, cache(4).Ra, cache(5).Ra));
-                break;
-              }
-              case DECAlpha::Mfc: {
-                state = State::Mfc_Function;
-                break;
-              }
-              case DECAlpha::Mbr: {
-                state = State::Mbr_Ra; // usually R26, R27 or R31
-                break;
-              }
-              case DECAlpha::Opr: {
-                state = State::Opr_Bit;
-                maps1[map_state(0u, State::Opr_Bit)].set_direct(op.relOpcode);
-                maps2[map_state(1u, State::Opr_Bit)].set(hash(op.Opcode, cache(1).Opcode, cache(1).Function));
-                break;
-              }
-              case DECAlpha::Pcd: {
-                state = State::Pcd_Function;
-                break;
-              }
-              case DECAlpha::Nop: {
-                state = State::Nop_Skip;
-                break;
-              }
-            }
-            count = 0u;
-          }
-          break;
-        }
-        case State::Bra_Ra: {
-          op.Ra += op.Ra + x.y;
-          if (count == 5u) {
-            state = State::Bra_Displacement;
-            count = 0u;
-            maps1[map_state(0u, State::Bra_Displacement)].set_direct(cache(1).Opcode);
-            maps2[map_state(1u, State::Bra_Displacement)].set_direct((cache(1).Opcode << 6u) | cache(2).Opcode);
-          }
-          break;
-        }
-        case State::Bra_Displacement: {
-          op.Displacement += op.Displacement + x.y;
-          if ((count % 7u) == 0) {
-            if (count < 21u) {
-              maps0[state].set(count / 7u);
-              maps1[map_state(0u, State::Bra_Displacement)].set(hash(count, cache(1).Opcode, op.Displacement) + 0x40u);
-              maps2[map_state(1u, State::Bra_Displacement)].set(hash(count, cache(1).Opcode, cache(2).Opcode, op.Displacement) + 0x1000u);
-            }
-            else {
-              state = State::OpCode;
-              count = 0u;
-            }
-          }
-          break;
-        }
-        case State::F_P_Function: {
-          op.Function = op.Function + x.y;
-          if (count == 11u) {
-            state = State::F_P_Ra;
-            count = 0u;
-            maps1[map_state(0u, State::F_P_Ra)].set_direct(cache(1).Rc);
-          }
-          else if ((count % 3u) == 0u) {
-            maps0[state].set(count / 3u);
-            maps1[map_state(0u, State::F_P_Function)].set(hash(count, op.Opcode, op.Function) + 0x40u);
-          }
-          break;
-        }
-        case State::F_P_Ra: {
-          op.Ra += op.Ra + x.y;
-          if (count == 5u) {
-            state = State::F_P_Rb;
-            count = 0u;
-            maps1[map_state(0u, State::F_P_Rb)].set_direct((op.Ra << 5u) | cache(1).Rc);
-          }
-          break;
-        }
-        case State::F_P_Rb: {
-          op.Rb += op.Rb + x.y;
-          if (count == 5u) {
-            state = State::F_P_Rc;
-            count = 0u;
-            maps1[map_state(0u, State::F_P_Rc)].set_direct((op.Ra << 5u) | op.Rb);
-          }
-          break;
-        }
-        case State::F_P_Rc: {
-          op.Rc += op.Rc + x.y;
-          if (count == 5u) {
-            lastRc = op.Rc;
-            state = State::OpCode;
-            count = 0u;
-          }
-          break;
-        }
-        case State::Mem_Ra: {
-          op.Ra += op.Ra + x.y;
-          if (count == 5u) {
-            state = State::Mem_Rb;
-            count = 0u;
-            maps1[map_state(0u, State::Mem_Rb)].set_direct(op.relOpcode);
-            maps2[map_state(1u, State::Mem_Rb)].set_direct((op.relOpcode << 5u) | cache(1).Rb);           
-            maps3[map_state(2u, State::Mem_Rb)].set_direct((op.relOpcode << 10u) | (cache(1).Rb << 5u) | cache(2).Rb);           
-            maps4[map_state(3u, State::Mem_Rb)].set(hash(op.Opcode, cache(1).Rb, cache(2).Rb, cache(3).Rb));
-            maps5[map_state(4u, State::Mem_Rb)].set(hash(op.Opcode, cache(1).Rb, cache(2).Ra, cache(3).Ra));
-            maps6[map_state(5u, State::Mem_Rb)].set(hash(op.Opcode, cache(1).Opcode, cache(1).Ra, cache(1).Rb));
-          }
-          break;
-        }
-        case State::Mem_Rb: {
-          op.Rb += op.Rb + x.y;
-          if (count == 5u) {
-            state = State::Mem_Displacement;
-            count = 0u;
-            maps1[map_state(0u, State::Mem_Displacement)].set_direct(op.Opcode);
-            maps2[map_state(1u, State::Mem_Displacement)].set(hash(op.Opcode, op.Rb));
-            maps3[map_state(2u, State::Mem_Displacement)].set(hash(op.Opcode, op.Rb, cache(1).Literal));
-          }
-          break;
-        }
-        case State::Mem_Displacement: {
-          op.Displacement += op.Displacement + x.y;
-          if ((count & 3u) == 0) {
-            if (count < 16u) {
-              maps0[state].set(count / 4u);
-              maps1[map_state(0u, State::Mem_Displacement)].set(hash(count, op.Opcode, op.Displacement));
-              maps2[map_state(1u, State::Mem_Displacement)].set(hash(count, op.Opcode, op.Rb, op.Displacement));
-              maps3[map_state(2u, State::Mem_Displacement)].set(hash(count, op.Opcode, op.Rb, cache(1).Literal, op.Displacement));
-            }
-            else {
-              state = State::OpCode;
-              count = 0u;
-            }
-          }
-          break;
-        }
-        case State::Mfc_Function: {
-          op.Function += op.Function + x.y;
-          if ((count & 3u) == 0) {
-            if (count < 16u)
-              maps0[state].set(count / 4u);
-            else {
-              state = State::Mfc_Ra; // usually R31
-              count = 0u;
-            }            
-          }
-          break;
-        }
-        case State::Mfc_Ra: {
-          op.Ra += op.Ra + x.y;
-          if (count == 5u) {
-            state = State::Mfc_Rb; // usually R31
-            count = 0u;
-          }
-          break;
-        }
-        case State::Mfc_Rb: {
-          op.Rb += op.Rb + x.y;
-          if (count == 5u) {
-            state = State::OpCode;
-            count = 0u;
-          }
-          break;
-        }
-        case State::Mbr_Ra: {
-          op.Ra += op.Ra + x.y;
-          if (count == 5u) {
-            state = State::Mbr_Rb;
-            count = 0u;
-            maps1[map_state(0u, State::Mbr_Rb)].set_direct(op.relOpcode);
-          }
-          break;
-        }
-        case State::Mbr_Rb: {
-          op.Rb += op.Rb + x.y;
-          if (count == 5u) {
-            state = State::Mbr_Displacement;
-            count = 0u;
-            maps1[map_state(0u, State::Mbr_Displacement)].set_direct(op.relOpcode);
-          }
-          break;
-        }
-        case State::Mbr_Displacement: {
-          op.Displacement += op.Displacement + x.y;
-          if ((count & 3u) == 0) {
-            if (count < 16u) {
-              maps0[state].set(count / 4u);
-              maps1[map_state(0u, State::Mbr_Displacement)].set_direct(((count >> 1u) & 0x6u) | op.relOpcode);
-            }
-            else {
-              state = State::OpCode;
-              count = 0u;
-            }
-          }
-          break;
-        }
-        case State::Opr_Bit: {
-          op.Bit = x.y;
-          state = State::Opr_Function;
-          count = 0u;
-          maps1[map_state(0u, State::Opr_Function)].set_direct((op.relOpcode << 1u) | op.Bit);
-          maps2[map_state(1u, State::Opr_Function)].set(hash(op.Opcode, op.Bit, cache(1).Opcode, cache(1).Function));
-          maps3[map_state(2u, State::Opr_Function)].set(hash(op.Opcode, op.Bit, cache(1).Opcode, cache(2).Opcode, cache(3).Opcode));
-          break;
-        }
-        case State::Opr_Function: {
-          op.Function += op.Function + x.y;
-          if (count == 7u) {
-            state = State::Opr_Ra;
-            count = 0u;
-            maps1[map_state(0u, State::Opr_Ra)].set_direct(cache(1).Ra);
-            maps2[map_state(1u, State::Opr_Ra)].set_direct(cache(1).Rc);
-            maps3[map_state(2u, State::Opr_Ra)].set_direct((cache(1).Ra << 5u) | cache(2).Ra);
-            maps4[map_state(3u, State::Opr_Ra)].set(hash(op.Bit, cache(1).Ra, cache(2).Ra, cache(3).Ra, cache(4).Ra));
-          }
-          break;
-        }
-        case State::Opr_Ra: {
-          op.Ra += op.Ra + x.y;
-          if (count == 5u) {
-            count = 0u;
-            if (op.Bit == 0u) {
-              state = State::Opr_Rb;
-              maps1[map_state(0u, State::Opr_Rb)].set_direct((cache(1).Ra << 5u) | cache(1).Rb);
-              maps2[map_state(1u, State::Opr_Rb)].set_direct((cache(1).Rb << 5u) | cache(2).Ra);
-              maps3[map_state(2u, State::Opr_Rb)].set_direct((op.Ra << 10u) | (cache(1).Ra << 5u) | cache(2).Ra);
-            }
-            else {
-              state = State::Opr_Literal;
-              maps1[map_state(0u, State::Opr_Literal)].set_direct((op.relOpcode << 7u) | op.Function);
-              maps2[map_state(1u, State::Opr_Literal)].set(hash(op.Opcode, op.Function, cache(1).Opcode, cache(1).Function));
-            }
-          }
-          break;
-        }
-        case State::Opr_Rb: {
-          op.Rb += op.Rb + x.y;
-          if (count == 5u) {
-            state = State::Opr_Unused;
-            count = 0u;
-          }
-          break;
-        }
-        case State::Opr_Unused: {
-          if (count == 3u) {
-            state = State::Opr_Rc;
-            count = 0u;
-            maps1[map_state(0u, State::Opr_Rc)].set_direct((op.Ra << 1u) | op.Bit);
-            maps2[map_state(1u, State::Opr_Rc)].set_direct((op.Ra << 10u) | (op.Rb << 5u) | cache(1).Ra);
-            maps3[map_state(2u, State::Opr_Rc)].set_direct((cache(1).Ra << 5u) | cache(2).Ra);
-            maps4[map_state(3u, State::Opr_Rc)].set_direct((cache(1).Ra << 10u) | (cache(2).Ra << 5u) | cache(3).Ra);
-          }
-          break;
-        }
-        case State::Opr_Literal: {
-          op.Literal += op.Literal + x.y;
-          if (count == 8u) {
-            state = State::Opr_Rc;
-            count = 0u;
-            maps1[map_state(0u, State::Opr_Rc)].set_direct((op.Ra << 1u) | op.Bit);
-            maps2[map_state(1u, State::Opr_Rc)].set_direct(0x8000u | (op.Ra << 10u) | (cache(1).Ra << 5u) | cache(1).Rc);
-            maps3[map_state(2u, State::Opr_Rc)].set_direct(0x400u | (cache(1).Ra << 5u) | cache(2).Ra);
-            maps4[map_state(3u, State::Opr_Rc)].set_direct(0x8000u | (cache(1).Ra << 10u) | (cache(2).Ra << 5u) | cache(3).Ra);
-          }
-          break;
-        }
-        case State::Opr_Rc: {
-          op.Rc += op.Rc + x.y;
-          if (count == 5u) {
-            lastRc = op.Rc;
-            state = State::OpCode;
-            count = 0u;
-          }
-          break;
-        }
-        case State::Pcd_Function:
-        case State::Nop_Skip: {
-          op.Function += op.Function + x.y;
-          if (count == 26u) {
-            state = State::OpCode;
-            count = 0u;
-          }
-          else if ((count % 3) == 0u)
-            maps0[state].set(count / 3u);
-          break;
-        }
-      }
-    }
-    if (count == 0u) {
-      maps0[state].set(0u);
-      if (state == State::OpCode) {
-        std::uint64_t ctx = hash(op.Opcode, op.Function);
-        maps1[map_state(0u, State::OpCode)].set(ctx);
-        maps2[map_state(1u, State::OpCode)].set(hash(ctx, cache(1).Opcode, cache(1).Function));
-        maps3[map_state(2u, State::OpCode)].set(ctx = hash(ctx, cache(1).Opcode, cache(2).Opcode));
-        maps4[map_state(3u, State::OpCode)].set(ctx = hash(ctx, cache(3).Opcode));
-        maps5[map_state(4u, State::OpCode)].set(hash(ctx, cache(4).Opcode));
-        maps6[map_state(5u, State::OpCode)].set(ctx = hash(op.Opcode, cache(1).Opcode, cache(2).Opcode, cache(3).Opcode, cache(4).Opcode, cache(5).Opcode));
-        maps7[map_state(6u, State::OpCode)].set(ctx = hash(ctx, cache(6).Opcode));
-        maps8[map_state(7u, State::OpCode)].set(ctx = hash(ctx, cache(7).Opcode));
-        maps9[map_state(8u, State::OpCode)].set(hash(op.Opcode, op.Ra, cache(1).Opcode, cache(1).Ra, cache(2).Opcode, cache(2).Ra));
-        maps10[map_state(9u, State::OpCode)].set(hash(op.Opcode, op.Rb, lastRc, cache(2).Opcode, cache(2).Rb, cache(3).Opcode, cache(3).Ra));
-        last[op.Format] = op;
-        cache.add(op);
-        op = { 0 };
-      }
-    }    
-   
-    maps0[state].mix1(m);
-    for (std::uint32_t i = 0u; i < nMaps - 1u; i++) {
-      if (((maps_mask[i] >> state) & 1u) != 0u)
-        maps[i][map_state(i, state)].mix1(m);
-    }
-
-    std::uint8_t const opcode = (state != State::OpCode) ? op.Opcode : cache(1).Opcode;
-
-    m.set(static_cast<std::uint32_t>(state) * 26u + count, State::Count * 26u);
-    m.set((state << 6u) | opcode, State::Count * 64u);
-    m.set( (hash(state, count, opcode))&2047, 2048u);
-    m.set( (hash(state, count, op.Opcode, cache(1).Opcode))&4095, 4096u);
-    m.set( (hash(state, count, cache(1).Opcode, cache(2).Opcode))&4095, 4096u);
-    m.set( (hash(state, count, cache(1).Opcode, cache(2).Opcode, cache(3).Opcode))&8191, 8192u);
-    m.set( (hash(state, count, op.Opcode, cache(1).Opcode, cache(2).Opcode, cache(3).Opcode))&8191, 8192u);
-    m.set( (hash(state, opcode, cache(1).Format, cache(2).Format, cache(3).Format, cache(4).Format))&8191, 8192u);
-    m.set( (hash(state, count, op.Opcode, op.Bit, op.Ra))&4095, 4096u);
-    m.set( (hash(state, op.Ra, op.Rb, op.Bit))&4095, 4096u);
-    m.set( (hash(state, op.Ra, last[DECAlpha::Mem].Ra, cache(1).Format))&4095, 4096u);
-x.DEC.state = state;
-    x.DEC.bcount = count;
-  return 1;
-}
- virtual ~decModel1(){}
-};
 
 
 #include "mod_ppmd.inc"
